careerpilot/requirements.txt
# Requirements
fastapi>=0.100.0
crewai>=0.30.0
python-dotenv>=1.0.0
reportlab>=4.0.0
pyyaml>=6.0
uvicorn>=0.23.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.0.4
google-api-python-client>=2.0.0